Lift normalization up the tree
